feature: render templates list and editing modal.
While implementing this feature I also moved the header and footer to
the main.tsx file because it was getting a little out of hand making
sure it is in the correct place on every single page component. I also
ran fixed some ui quirks that popped up on the builder and templates
page like content overflow in the x direction. I just need to finish the
functional aspects of the template editing modal and it show be good to
go.