feature: templte tags
Only thing left to do at this point is make it so that she can delete
tags. Also I created an entire tag file in back end but she honestly
might not need this so I may just change this to only show the tags on
her templates. but for now I will just do both implementations bc why
not.